perf: split homepage into server and client components (#201)
