+++ apps/web/components/home/index.ts
@@ -1,0 +1,6 @@
+export { default as HeroSection } from './HeroSection';
+export { default as FeaturesSection } from './FeaturesSection';
+export { default as HowItWorksSection } from './HowItWorksSection';
+export { default as BetaCtaSection } from './BetaCtaSection';
+export { default as FaqSection } from './FaqSection';
+export { default as CtaSection } from './CtaSection';
